moved central text logic to the component
