+++ packages/react-ui/src/widgets/PieWidgetUI/components/PieCentralText.js
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import { Box, styled } from '@mui/material';
 import Typography from '../../../components/atoms/Typography';
+import { calculatePercentage, findLargestCategory } from '../../utils/chartUtils';
 
 const Root = styled(Box)(({ theme }) => ({
   zIndex: 1,
@@ -28,10 +29,50 @@ const MarkerColor = styled(Box)(({ theme }) => ({
   height: theme.spacing(1)
 }));
 
-function PieCentralText({ item }) {
-  const { name, percentage, color } = item;
+function PieCentralText({ data, selectedCategories }) {
+  console.log('data', data);
+  const [selectedItem, setSelectedItem] = useState({});
 
-  if (!item) {
+  // Select the largest category to display in CentralText and calculate its percentage from the total
+  const topSelectedCategory = useMemo(() => {
+    if (!data || data.length === 0) {
+      return null;
+    }
+
+    let array;
+    if (selectedCategories.length > 0) {
+      array = data.filter((dataItem) => selectedCategories.includes(dataItem.name));
+    } else {
+      array = data;
+    }
+
+    const largestCategory = findLargestCategory(array);
+    const category = array.find((element) => element === largestCategory);
+
+    if (!category) {
+      return null;
+    }
+
+    let sumValue = 0;
+    for (const category of data) {
+      sumValue += category.value;
+    }
+
+    const percentage = calculatePercentage(category.value, sumValue);
+    category.percentage = percentage;
+
+    return category;
+  }, [data, selectedCategories]);
+
+  useEffect(() => {
+    if (topSelectedCategory) {
+      setSelectedItem(topSelectedCategory);
+    }
+  }, [topSelectedCategory, setSelectedItem]);
+
+  const { name, percentage, color } = selectedItem;
+
+  if (!selectedItem) {
     return null;
   }
 
@@ -49,11 +90,13 @@ function PieCentralText({ item }) {
 }
 
 PieCentralText.propTypes = {
-  item: PropTypes.shape({
-    name: PropTypes.string,
-    percentage: PropTypes.string,
-    color: PropTypes.string
-  })
+  data: PropTypes.arrayOf(
+    PropTypes.shape({
+      name: PropTypes.string,
+      value: PropTypes.number,
+      color: PropTypes.string
+    })
+  )
 };
 
 export default PieCentralText;
